requirements.txt
fastapi
uvicorn
pytest
requests
allure-pytest
httpx